+++ Código/Classes/Business/FFEV.java
@@ -2,7 +2,7 @@ public class FFEV implements IFidelidade {
     public static final float VALOR_DESC = 0.2F;
 
     @Override
-    public float calcularDesconto(Pedido pedido) {
+    public double calcularDesconto(Pedido pedido) {
         return pedido.valorTotal() * (1 - VALOR_DESC);
     }
 }
