+++ Código/Classes/Business/FPreto.java
@@ -2,7 +2,7 @@ public class FPreto implements IFidelidade {
     public static final float VALOR_DESC = 0.1F;
 
     @Override
-    public float calcularDesconto(Pedido pedido) {
+    public double calcularDesconto(Pedido pedido) {
         return pedido.valorTotal() * (1 - VALOR_DESC);
     }
 }
